refactor(search): restructure module and add recent food store
