+++ src/modules/search/ui/TemplateSearchBar.tsx
