+++ src/modules/search/ui/TemplateSearchResultItem.tsx
