+++ src/modules/search/ui/TemplateSearchResults.tsx
@@ -1,13 +1,13 @@
 import { type Accessor, For, Show } from 'solid-js'
 
 import { type Template } from '~/modules/diet/template/domain/template'
+import { SearchLoadingIndicator } from '~/modules/search/ui/SearchLoadingIndicator'
+import { TemplateSearchResultItem } from '~/modules/search/ui/TemplateSearchResultItem'
 import {
   debouncedTab,
   templates,
 } from '~/modules/template-search/application/usecases/templateSearchState'
 import { Alert } from '~/sections/common/components/Alert'
-import { SearchLoadingIndicator } from '~/sections/search/components/SearchLoadingIndicator'
-import { TemplateSearchResultItem } from '~/sections/search/components/TemplateSearchResultItem'
 
 export function TemplateSearchResults(props: {
   search: string
