+++ src/modules/search/ui/TemplateSearchTabs.tsx
